+++ R/md_stock_symbol.R
@@ -2,7 +2,7 @@
 #' @import data.table
 #' @importFrom stringi stri_unescape_unicode
 md_stk_syb_hk = function(source="eastmoney", return_price=TRUE) {
-  symbol = exchange = market_symbols = stock_list_hk2 = ticktime = f1 = name = NULL
+  symbol = exchange = market_symbols = stock_list_hk2 = ticktime = f1 = name = market = mktcode = sybcode = NULL
 
   if (source=="sina") {
     url_hk_sina <- c(
@@ -15,58 +15,61 @@ md_stk_syb_hk = function(source="eastmoney", return_price=TRUE) {
     setnames(stock_list_hk, c('engname', 'lasttrade', 'prevclose', 'pricechange', 'changepercent', 'market_value', 'pe_ratio'), c('name_eng', 'close', 'close_prev', 'change', 'change_pct', 'cap_total', 'pe'))
   } else if (source == 'eastmoney') {
         fid = 
-            c(1, 12, 14, 124, 17, 15, 16, 2, 5, 6, 8)
+            c(1, 12, 14, 124, 17, 15, 16, 2, 5, 6, 8, 29, 62, 23, 20, 21)
             # setdiff(1:149,c(36:61,100:108, 112:123,128:138))
+        # "f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f11,f12,f13,f14,f15,f16,f17,f18,f19,f20,f21,f22,f23,f24,f25,f62,f128,f136,f115,f152", 
         
         url_hk_em = sprintf(
             "http://72.push2.eastmoney.com/api/qt/clist/get?pn=1&pz=50000&po=1&np=1&ut=%s&fltt=2&invt=2&fid=f3&fs=%s&fields=%s&_=%s", 
             'bd1d9ddb04089700cf9c27f6f7426281', 
             "m:128%20t:1", 
             paste0('f', fid, collapse = ','),
-            # "f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f11,f12,f13,f14,f15,f16,f17,f18,f19,f20,f21,f22,f23,f24,f25,f62,f128,f136,f115,f152", 
             date_num(Sys.time(), 'ms')
             ) # m:128%20t:1,m:128%20t:2,m:128%20t:3,m:128%20t:4,
-        datem = read_apidata_eastmoney(url_hk_em, type = 'real_cn')
         
-        setnames(datem, 
-                 c('f12', 'f14', 'f124', 'f17', 'f15', 'f16', 'f2', 'f5', 'f6', 'f8'), 
-                 c('symbol', 'name', 'ticktime', 'open', 'high', 'low', 'close', 'volume', 'amount', 'turnover' ))
+        datem = read_apidata_eastmoney(url_hk_em, type = 'real_cn')
+        setnames(datem, fidnam('fi'), fidnam('ni'), skip_absent=TRUE)
+        
         datem = datem[
             , ticktime := as.POSIXct(ticktime, origin = '1970-01-01')
-        ][grepl('牛|熊', name), market := 'bullbear'
-        ][grepl('沽|购', name), market := 'warrant']
+        ][mktcode == 1024, market := 'bullbear'
+        ][mktcode == 64, market := 'warrant']
+        
         stock_list_hk = datem[f1==3] 
     }
 
 
   stock_list_hk = stock_list_hk[,`:=`(
-      symbol = paste0(symbol, '.HK'),
+      symbol = paste0(sybcode, '.HK'),
       date = as_date(ticktime)
-  )][, c('symbol', 'name', 'date', 'open', 'high', 'low', 'close', 'volume', 'amount', 'turnover')]
+  )][is.na(market)]
   
   return(stock_list_hk)
 }
 
 md_stk_syb_us = function(exchange) {
-  . = symbol = name = sector = industry = country = ipoyear = marketCap = NULL
+  . = symbol = name = sector = industry = country = ipoyear = marketCap = ticktime = sybcode = NULL
   # c("AMEX", "NASDAQ", "NYSE")
   exchange = toupper(exchange)
   exchange2code = list(NASDAQ = '105', NYSE='106', AMEX = '107')
+  fid = c(1, 12, 14, 124, 17, 15, 16, 2, 5, 6, 8, 29, 62, 23, 20, 21)
+  # "f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f12,f13,f14,f15,f16,f17,f18,f20,f21,f23,f24,f25,f26,f22,f33,f11,f62,f128,f136,f115,f152", 
   
   url = sprintf(
       "http://72.push2.eastmoney.com/api/qt/clist/get?pn=1&pz=20000&po=1&np=1&ut=bd1d9ddb04089700cf9c27f6f7426281&fltt=2&invt=2&fid=f3&fs=m:%s&fields=%s&_=%s", 
       exchange2code[[exchange]], 
-      paste0('f', c(1, 12, 14, 124, 17, 15, 16, 2, 5, 6, 8), collapse = ','),
-      # "f1,f2,f3,f4,f5,f6,f7,f8,f9,f10,f12,f13,f14,f15,f16,f17,f18,f20,f21,f23,f24,f25,f26,f22,f33,f11,f62,f128,f136,f115,f152", 
+      paste0('f', fid, collapse = ','),
       date_num(Sys.time(), 'ms')
   )
   
   datem = read_apidata_eastmoney(url, type = 'real_us')
-  setnames(datem, 
-           c('f12', 'f14', 'f124', 'f17', 'f15', 'f16', 'f2', 'f5', 'f6', 'f8'), 
-           c('symbol', 'name', 'ticktime', 'open', 'high', 'low', 'close', 'volume', 'amount', 'turnover' ))
+  setnames(datem, fidnam('fi'), fidnam('ni'), skip_absent=TRUE)
+  
   datem = datem[, ticktime := as.POSIXct(ticktime, origin = '1970-01-01')
-              ][, date := as_date(ticktime)] 
+              ][, `:=`(
+                  date = as_date(ticktime), 
+                  symbol = sybcode
+              )] 
   
   if (FALSE) {
       url = sprintf(
@@ -88,7 +91,7 @@ md_stk_syb_us = function(exchange) {
 md_stock_symbol_exchange = function(XCHG=NULL, print_step=1L) {
   exchange = NULL
   
-  exchange_list = c('sse','szse', 'bse', 'hkex', "amex", "nasdaq", "nyse")
+  exchange_list = c('sse','szse', 'hkex', "amex", "nasdaq", "nyse")
   while ((is.null(XCHG) || length(XCHG)==0)) {
     XCHG = select_rows_df(dt = setDT(list(exchange = exchange_list)), column = 'exchange')[,exchange]
   }
@@ -98,7 +101,7 @@ md_stock_symbol_exchange = function(XCHG=NULL, print_step=1L) {
   i = 1
   exc_len = length(XCHG)
   dat_lst = NULL
-  if (any(c("sse", "szse", 'bse') %in% XCHG)) {
+  if (any(c("sse", "szse") %in% XCHG)) {
     dtmp = md_stocka_eastmoney()
     temp = rbindlist(dtmp[c('stock','index')])[exchange %in% XCHG]
     exc = intersect(c("sse", "szse", 'bse'), XCHG)
@@ -152,7 +155,7 @@ md_stock_symbol_exchange = function(XCHG=NULL, print_step=1L) {
 md_stock_symbol = function(exchange=NULL, ...) {
     datlst = md_stock_symbol_exchange(exchange)
     datlst2 = lapply(datlst, function(x) {
-        cols = c('symbol', 'name', 'date', 'open', 'high', 'low', 'close', 'volume', 'amount', 'turnover', 'market')
+        cols = c('symbol', 'name', 'date', 'open', 'high', 'low', 'close', 'volume', 'amount', 'turnover', 'market', 'mktcode', 'pe_ttm', 'pb', 'cap_total', 'cap_market')
         
         x[, intersect(cols, names(x)), with = FALSE]
     } )